スタンプを押した人の場所からユーザーモーダルを開けるように close #687
